+++ src/codex_issue_memory/learning/__init__.py
@@ -1,11 +1,12 @@
 from .posteriors import BetaPosterior, build_beta_posterior, deterministic_rng, shrinkage_weight
-from .safe_override import SafeOverridePolicy, SafeOverrideResult
+from .safe_override import FP_SAFETY_BLOCK_THRESHOLD, SafeOverridePolicy, SafeOverrideResult
 from .strategy_bandit import StrategyBanditOutcome, StrategyThompsonBandit
 
 __all__ = [
     "BetaPosterior",
     "build_beta_posterior",
     "deterministic_rng",
+    "FP_SAFETY_BLOCK_THRESHOLD",
     "shrinkage_weight",
     "SafeOverridePolicy",
     "SafeOverrideResult",